new class seems to be working
